+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0.8 - under development
 
 - Added `davidhirtz\yii2\media\models\forms\TransformationForm`
+- Created an index on `path` column in `folder` table
 
 ## 2.0.7 (Nov 14, 2023)